address-book 프로젝트에 interface modules 적용
