+++ 2_address-book/src/types.ts
@@ -1,0 +1,19 @@
+interface PhoneNumberDictionary {
+    [phone: string]: {
+      num: number;
+    };
+  }
+  
+  interface Contact {
+    name: string;
+    address: string;
+    phones: PhoneNumberDictionary;
+  }
+  
+  enum PhoneType {
+    Home = 'home',
+    Office = 'office',
+    Studio = 'studio',
+  }
+
+  export {Contact, PhoneType};
